pyproject.toml
[tool.poetry]
name = "contribution-complexity"
version = "0.1.0"
description = " Tool to compute the complexity of a Git contributions "
authors = ["HelgeCPH <ropf@itu.dk>"]
readme = "README.md"
repository = "https://github.com/HelgeCPH/contribution-complexity"
documentation = "https://github.com/HelgeCPH/contribution-complexity/blob/main/README.md"
classifiers = [
    "Development Status :: 4 - Beta", 
    "Environment :: Console", 
    "Intended Audience :: Developers", 
    "Intended Audience :: Education", 
    "Intended Audience :: Science/Research", 
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)", 
    "Programming Language :: Python :: 3", 
    "Topic :: Software Development", 
    "Topic :: Utilities",
]
exclude = ["paper/", "data/", "experiment/"]

[tool.poetry.dependencies]
python = "^3.9"
docopt = "^0.6.2"
PyDriller = "^1.15.5"
toml = "^0.10.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.22.0"
pandas = "^1.2.3"
PyGithub = "^1.54.1"
matplotlib = "^3.4.1"
scipy = { version = "^1.6.2", python = ">=3.9,<3.10" }
tabulate = "^0.8.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
contribcompl = "contribution_complexity.compute:run"